Handle Windows paths in module import
Running this on Windows did not work with the replace of /
This imports the file directly from: https://docs.python.org/3/library/importlib.html#importing-a-source-file-directly

This could allow eventually for the anonymizer file to be located anywhere instead of the specific anonymizer directory. eg. app/anonymizer.py allowing to keep everything together for each app.